custom-elements/instantiating-custom-elements/extensions-to-document-interface/created-callback-create-element.html

<!DOCTYPE html>
<html>
<head>
<title>Created callback should be called if createElement is called</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="Created callback should be called if createElement is called">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#types-of-callbacks">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
test(function() {
    var doc = newHTMLDocument();
    var callbackCalled = false;
    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function() {
        callbackCalled = true;
    };
    var name = 'x-a';
    doc.registerElement(name, {prototype: proto});
    var customElement = doc.createElement(name);
    assert_true(callbackCalled, 'Callback created should be called');
}, 'Created callback should be called if createElement is called. Type is not specified');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalled = false;
    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function() {
        callbackCalled = true;
    };
    var name = 'x-b';
    doc.registerElement(name, {prototype: proto});
    var customElement = doc.createElement(name, name);
    assert_true(callbackCalled, 'Callback created should be called');
}, 'Created callback should be called if createElement is called. Type is specified');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalled = false;
    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function() {
        callbackCalled = true;
    };
    var name = 'x-c';
    var customElement = doc.createElement(name);
    assert_false(callbackCalled, 'Callback created should not be called');
    doc.registerElement(name, {prototype: proto});
    assert_true(callbackCalled, 'Callback created should be called');
}, 'Created callback should be called if createElement is called. ' +
    'Element is unresolved at first');


test(function() {
    var doc = newHTMLDocument();
    HTML5_ELEMENTS.forEach(function(value) {
        var obj = doc.createElement(value);
        var name = 'x-d-' + value;
        var proto = Object.create(obj.constructor.prototype);
        proto.createdCallback = function() {
            callbackCalled = true;
        };    
        doc.registerElement(name, {prototype: proto, extends: value});
        var customElement = doc.createElement(value, name);
        assert_true(callbackCalled, 'Callback created should be called');
    });
}, 'Created callback should be called if createElement is called. Test different types');
</script>
</body>
</html>
